eventapp/events/paginators.py
from rest_framework.pagination import PageNumberPagination

class EventPaginator(PageNumberPagination):
    page_size = 5
    page_query_param = 'page'
    page_size_query_param = 'page_size'

class CommentPaginator(PageNumberPagination):
    page_size = 10
    page_query_param = 'page'
    page_size_query_param = 'page_size'